+++ provider/pkg/await/ingress.go
@@ -24,7 +24,6 @@ import (
 	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/await/informers"
 	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/clients"
 	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/kinds"
-	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/metadata"
 	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/openapi"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
@@ -116,7 +115,7 @@ func (iia *ingressInitAwaiter) Await() error {
 	defer close(stopper)
 
 	informerFactory := informers.NewInformerFactory(iia.config.clientSet,
-		informers.WithNamespaceOrDefault(iia.config.currentInputs.GetNamespace()))
+		informers.WithNamespaceOrDefault(iia.config.currentOutputs.GetNamespace()))
 	informerFactory.Start(stopper)
 
 	ingressEvents := make(chan watch.Event)
@@ -144,7 +143,7 @@ func (iia *ingressInitAwaiter) Await() error {
 	}
 	go serviceInformer.Informer().Run(stopper)
 
-	timeout := metadata.TimeoutDuration(iia.config.timeout, iia.config.currentInputs, DefaultIngressTimeoutMins*60)
+	timeout := iia.config.getTimeout(DefaultIngressTimeoutMins * 60)
 	return iia.await(ingressEvents, serviceEvents, endpointsEvents, make(chan struct{}), time.After(60*time.Second), time.After(timeout))
 }
 
@@ -155,7 +154,7 @@ func (iia *ingressInitAwaiter) Read() error {
 	}
 
 	// Get live versions of Ingress.
-	ingress, err := ingressClient.Get(iia.config.ctx, iia.config.currentInputs.GetName(), metav1.GetOptions{})
+	ingress, err := ingressClient.Get(iia.config.ctx, iia.config.currentOutputs.GetName(), metav1.GetOptions{})
 	if err != nil {
 		// IMPORTANT: Do not wrap this error! If this is a 404, the provider need to know so that it
 		// can mark the deployment as having been deleted.
@@ -288,7 +287,7 @@ func (iia *ingressInitAwaiter) processServiceEvent(event watch.Event) {
 }
 
 func (iia *ingressInitAwaiter) processIngressEvent(event watch.Event) {
-	inputIngressName := iia.config.currentInputs.GetName()
+	inputIngressName := iia.config.currentOutputs.GetName()
 
 	ingress, isUnstructured := event.Object.(*unstructured.Unstructured)
 	if !isUnstructured {
@@ -509,25 +508,25 @@ func (iia *ingressInitAwaiter) makeClients() (
 	ingressClient, endpointsClient, servicesClient dynamic.ResourceInterface, err error,
 ) {
 	ingressClient, err = clients.ResourceClient(
-		kinds.Ingress, iia.config.currentInputs.GetNamespace(), iia.config.clientSet)
+		kinds.Ingress, iia.config.currentOutputs.GetNamespace(), iia.config.clientSet)
 	if err != nil {
 		return nil, nil, nil, errors.Wrapf(err,
 			"Could not make client to watch Ingress %q",
-			iia.config.currentInputs.GetName())
+			iia.config.currentOutputs.GetName())
 	}
 	endpointsClient, err = clients.ResourceClient(
-		kinds.Endpoints, iia.config.currentInputs.GetNamespace(), iia.config.clientSet)
+		kinds.Endpoints, iia.config.currentOutputs.GetNamespace(), iia.config.clientSet)
 	if err != nil {
 		return nil, nil, nil, errors.Wrapf(err,
 			"Could not make client to watch Endpoints associated with Ingress %q",
-			iia.config.currentInputs.GetName())
+			iia.config.currentOutputs.GetName())
 	}
 	servicesClient, err = clients.ResourceClient(
-		kinds.Service, iia.config.currentInputs.GetNamespace(), iia.config.clientSet)
+		kinds.Service, iia.config.currentOutputs.GetNamespace(), iia.config.clientSet)
 	if err != nil {
 		return nil, nil, nil, errors.Wrapf(err,
 			"Could not make client to watch Services associated with Ingress %q",
-			iia.config.currentInputs.GetName())
+			iia.config.currentOutputs.GetName())
 	}
 
 	return